+++ packages/core/src/js/tools/sentryMetroSerializer.ts
@@ -1,18 +1,11 @@
 import * as crypto from 'crypto';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import type { MixedOutput, Module, ReadOnlyGraph } from 'metro';
-import type * as countLinesType from 'metro/private/lib/countLines';
 
 import type { Bundle, MetroSerializer, MetroSerializerOutput, SerializedBundle, VirtualJSOutput } from './utils';
 import { createDebugIdSnippet, createSet, determineDebugIdFromBundleSource, stringToUUID } from './utils';
+import countLines from './vendor/metro/countLines';
 import { createDefaultMetroSerializer } from './vendor/metro/utils';
-
-let countLines: typeof countLinesType;
-try {
-  countLines = require('metro/private/lib/countLines');
-} catch (e) {
-  countLines = require('metro/src/lib/countLines');
-}
 
 type SourceMap = Record<string, unknown>;